# cmd/docs/swagger.yaml
basePath: /api
definitions:
  structure.Song:
    properties:
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      release_date:
        type: string
      song:
        type: string
      song_id:
        type: integer
      text:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: qwaq.dev@gmail.com
    name: QwaQ
  description: Онлайн библиотека песен
  title: Songs API
  version: "1.0"
paths:
  /api/song/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет песню из базы данных по её уникальному идентификатору.
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Песня успешно удалена
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Песня не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удаление песни по ID
      tags:
      - Songs
    get:
      consumes:
      - application/json
      description: Возвращает информацию о песне по её ID, включая название и группу.
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные о песне
          schema:
            $ref: '#/definitions/structure.Song'
        "400":
          description: Некорректный запрос (например, ID не является числом)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Песня не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение информации о песне
      tags:
      - Songs
    patch:
      consumes:
      - application/json
      description: Позволяет обновить одно или несколько полей песни по её ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления (только изменяемые поля)
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Песня успешно обновлена
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Песня не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Частичное обновление песни
      tags:
      - Songs
    put:
      consumes:
      - application/json
      description: Обновляет информацию о песне по её ID, включая название и группу.
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - description: Объект с обновлёнными данными
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/structure.Song'
      produces:
      - application/json
      responses:
        "200":
          description: Песня успешно обновлена
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный запрос или ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Песня не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление данных о песне
      tags:
      - Songs
  /api/song/{id}/text:
    get:
      consumes:
      - application/json
      description: Возвращает текст песни по ID с возможностью указать номер страницы
        и количество строк на странице.
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 2
        description: Количество строк текста на странице
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Текст песни с пагинацией
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Песня или текст не найдены
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение текста песни с пагинацией
      tags:
      - Songs
  /api/songs:
    get:
      consumes:
      - application/json
      description: Возвращает список песен с возможностью фильтрации по названию и
        группе, а также с пагинацией.
      parameters:
      - description: Фильтр по названию песни (поиск по подстроке)
        in: query
        name: song
        type: string
      - description: Фильтр по названию группы (поиск по подстроке)
        in: query
        name: group
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 10
        description: Количество записей на странице
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список песен
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение списка песен
      tags:
      - Songs
    post:
      consumes:
      - application/json
      description: Позволяет добавить песню с указанием названия и группы.
      parameters:
      - description: Данные песни (название и группа)
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/structure.Song'
      produces:
      - application/json
      responses:
        "201":
          description: Песня успешно добавлена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавление новой песни
      tags:
      - Songs
swagger: "2.0"